// src/features/auth-register/schema.ts
import { z } from 'zod'

export const registerSchema = z.object({
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters long"),
  firstName: z.string().min(1, "First name is required"),
  lastName: z.string().min(1, "Last name is required"),
})

export type RegisterInput = z.infer<typeof registerSchema>